Power Automate Integration

+++ businessCentral/src/ADLSEClearTrackedDeletions.Codeunit.al
@@ -28,7 +28,8 @@ codeunit 82573 "ADLSE Clear Tracked Deletions"
             repeat
                 ADLSEDeletedRecord.SetRange("Table ID", ADLSETable."Table ID");
                 ADLSEDeletedRecord.SetFilter("Entry No.", '<=%1', ADLSETableLastTimestamp.GetDeletedLastEntryNo(ADLSETable."Table ID"));
-                ADLSEDeletedRecord.DeleteAll();
+                if not ADLSEDeletedRecord.IsEmpty() then
+                    ADLSEDeletedRecord.DeleteAll();
 
                 ADLSETableLastTimestamp.SaveDeletedLastEntryNo(ADLSETable."Table ID", 0);
             until ADLSETable.Next() = 0;
